added check for zone existance so its not duplicated, causing bind9 failing to start

+++ src/start.sh
@@ -4,10 +4,12 @@
 
 IFS='|' read -ra ADDR <<< "$DOMAIN"
 for i in "${ADDR[@]}"; do
-    cat /template/update.php | sed "s/<DOMAIN>/${i}/g" > /var/www/html/$i.php
-    cat /template/zone | sed "s/<DOMAIN>/${i}/g" > /var/lib/bind/d.$i
-    cat /template/named.conf.local.zone_config | sed "s/<DOMAIN>/${i}/" >> /etc/bind/named.conf.local
-    echo "" >> /etc/bind/named.conf.local
+    if [ ! -f /var/lib/bind/d.$i ]; then
+        cat /template/zone | sed "s/<DOMAIN>/${i}/g" > /var/lib/bind/d.$i
+        cat /template/update.php | sed "s/<DOMAIN>/${i}/g" > /var/www/html/$i.php
+        cat /template/named.conf.local.zone_config | sed "s/<DOMAIN>/${i}/" >> /etc/bind/named.conf.local
+        echo "" >> /etc/bind/named.conf.local
+    fi
 done
